Clean useless templates
Remove tracker views

Remove tracker tests

Move badly located email tests

Remove tracker middleware
