--- lemarche/templates/includes/_header_notice_migration.html
@@ -1,6 +1,0 @@
-<div id="header-notice" class="alert alert-info fr-p-3 text-center">
-    <p class="mb-0">
-        Le marché de l'inclusion <a href="{% url 'pages:2021-10-06-le-marche-fait-peau-neuve' %}">fait peau neuve</a> !
-        N'hésitez pas à <a href="{% url 'pages:contact' %}">nous remonter</a> le moindre problème.
-    </p>
-</div>
