--- lemarche/templates/includes/_messages.html
@@ -1,14 +1,0 @@
-{% load static %}
-{% for message in messages %}
-    {% if message.extra_tags and message.extra_tags == "modal_message_bizdev" %}
-        {% include "includes/_messages_modal_bizdev.html" with index=forloop.counter message=message %}
-    {% else %}
-        <div class="alert {{ message.tags }} alert-dismissible fade show mt-3"
-             role="alert">
-            <button type="button" class="close" data-dismiss="alert" aria-label="Fermer">
-                <i class="ri-close-line"></i>
-            </button>
-            {{ message | safe }}
-        </div>
-    {% endif %}
-{% endfor %}
