--- lemarche/templates/includes/_messages_modal_bizdev.html
@@ -1,34 +1,0 @@
-{% load static %}
-<div class="modal modal-message-alert"
-     id="message{{ index }}"
-     tabindex="-1"
-     aria-labelledby="{{ message }}"
-     style="display: none"
-     aria-hidden="true">
-    <div class="modal-dialog modal-lg modal-dialog-centered">
-        <div class="modal-content">
-            <div class="modal-header">
-                {% comment %} <h3 class="modal-title">Information importante</h3> {% endcomment %}
-                <button type="button" class="close" data-dismiss="modal" aria-label="Fermer">
-                    <i class="ri-close-line"></i>
-                </button>
-            </div>
-            <div class="modal-body">
-                <div class="fr-grid-row">
-                    <div class="fr-col-md-4 d-flex align-items-center justify-content-center">
-                        <img src="{% static 'images/bizdev-icon.jpg' %}"
-                             class="img-fluid rounded-circle border border-primary shadow"
-                             alt="Image du chargé de votre besoin"
-                             loading="lazy"/>
-                    </div>
-                    <div class="fr-col-md-8">{{ message | safe }}</div>
-                </div>
-            </div>
-            <div class="modal-footer">
-                <button type="button"
-                        class="btn btn-sm btn-outline-primary"
-                        data-dismiss="modal">Fermer</button>
-            </div>
-        </div>
-    </div>
-</div>
