--- lemarche/templates/includes/_newsletter_button.html
@@ -1,6 +1,0 @@
-<a href="#" data-token="0d87cc6eac81c8037144017a95fdecdb" onclick="mjOpenPopin(event, this)" class="btn btn-link pr-0 text-decoration-none {{ custom_class }}" rel="noopener">
-    <i class="ri-mail-send-line ri-lg"></i>
-    <span class="ml-1 font-weight-bold">
-        Newsletter achat
-    </span>
-</a>
