--- lemarche/templates/includes/_tracker_itou.html
@@ -1,82 +1,0 @@
-<script type="text/javascript">
-async function track(page, action, meta={}) {
-    const data = computeRequestBody(page, action, meta);
-
-    if (typeof window !== 'undefined') {
-        await fetch(`${window.location.origin}/track/`, {
-            method: 'POST',
-            headers: {
-                'X-CSRFToken': '{{ csrf_token }}',
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(data)
-        });
-    }
-}
-
-function computeRequestBody(page, action, metaIn) {
-    const timeNow = new Date();
-    // Optionnaly add meta data (in key/value pairs) from url query string
-    const meta = Object.assign(metaIn, getUrlMeta());
-    // user info
-    meta['is_admin'] = {% if user.is_authenticated %}{% if user.is_admin %}true{% else %}false{% endif %}{% else %}false{% endif %};
-    meta['user_kind'] = {% if user.is_authenticated %}'{{ user.kind }}'{% else %}null{% endif %};
-    meta['user_id'] = {% if user.is_authenticated %}{{ user.id }}{% else %}null{% endif %};
-    meta['siae_id'] = {% if siae %}{{ siae.id }}{% else %}null{% endif %};
-    // default data
-    meta['source'] = 'bitoubi_frontend';
-
-    return {
-        page: page,
-        action: action,
-        meta: meta,
-    }
-}
-
-
-function getUrlMeta() {
-    // Avoid using JSON and interpreting user-provided values
-    var meta = {};
-    const doc = typeof document !== `undefined` ? document : null;
-    try {
-        const check = (new URL(doc.location)).searchParams.get('t');
-        const split1 = check.split(';');
-        for (var tuple of split1) {
-            const keyval = tuple.split(':');
-            const key = keyval[0].replace(/[^\-_0-9a-zA-Z]/g,'');
-            const val = keyval[1].replace(/[^\-_0-9a-zA-Z]/g,'');
-            meta[key] = val;
-        }
-    } finally {
-        return meta;
-    }
-}
-
-// track each link & button click
-document.addEventListener('click', async (event) => {
-    var event = event || window.event;
-    var target = event.target || event.srcElement;
-
-    var loops = 3;
-    var meta = {};
-    while (target && loops > 0) {
-        if (target instanceof HTMLAnchorElement || target instanceof HTMLButtonElement) {
-            meta = {
-                'href': target.getAttribute('href'),
-                'id' : target.id,
-                'class': target.className
-            }
-            // stop loop
-            loops = 0;
-            // send track
-            await track(
-                page=window.location.pathname,
-                action='click',
-                meta=meta
-            );
-        }
-        target = target.parentNode;
-        loops -= 1;
-    }
-}, false);
-</script>
