--- lemarche/templates/includes/forms/_bootstrap_field_radio.html
@@ -1,24 +1,0 @@
-{% comment %}
-    takes two arguments : fields and form_group_class
-{% endcomment %}
-
-<div class="{% if form_group_class %}{{ form_group_class }}{% else %}form-group{% endif %}">
-    {{ form_field.label_tag }}
-    <ul class="list-unstyled mb-1">
-        {% for pk, choice in form_field.field.widget.choices %}
-        <li>
-            <div class="custom-control custom-radio">
-                <input id="id_{{ form_field.name }}_{{ forloop.counter0 }}"
-                    class="custom-control-input {% if form_field.errors %} is-invalid{% endif %}"
-                    name="{{ form_field.name }}" type="{{ form_field.field.widget.input_type }}" value="{{ pk }}" />
-                <label for="id_{{ form_field.name }}_{{ forloop.counter0 }}" class="custom-control-label">{{ choice }}</label>
-                {% if forloop.last and form_field.errors %}
-                    {% for error in form_field.errors %}
-                        <div class="invalid-feedback">{{ error }}</div>
-                    {% endfor %}
-                {% endif %}
-            </div>
-        </li>
-        {% endfor %}
-    </ul>
-</div>
