+++ lemarche/templates/layouts/base.html
@@ -109,7 +109,6 @@
     <script type="text/javascript" src="{% static 'js/dropdown_menu.js' %}"></script>
     {% if BITOUBI_ENV not in "dev" %}
       {% include "includes/_tracker_tarteaucitron.html" %}
-      {% include "includes/_tracker_itou.html" %}
     {% endif %}
     {% block extra_js %}
     {% endblock extra_js %}
